Fix Dungeon_Tiles_Sprites tileset
There was some problem using the extruded version of the tiles file (or
possibly just because it was the same image as the background tiles) for
sprites; only the first 1320 tiles were being registered. I switched the
tileset to use the unextruded image file for the sprites and now things
seem to work correctly.