Change player sprite x origin to 0
This prevents awkwardly shifting the position of the player's hitbox
when changing directions which happens because the sprite size is
different for different directons. When the x origin is 0.5, the change
in sprite width can cause a collision against terrain that forces the
player to "jump" to the edge of that terrain.