Change save data to use room relative player position
IMPORTANT: this is a breaking change for saved games! Previously saved
player position will be lost and they will be returned to the start.
However, all other saved data will remain intact so this should be ok.

This is necessary so we can expand the map without risking further saved
game breakage (as long as the rooms with the saved position do not
change too much).